[AN-4512] Fixes emoji visibility issue on tablet and filtering of unsupported emojis Ticket: https://wearezeta.atlassian.net/browse/AN-4512
