+++ wire-ui/src/main/java/com/waz/zclient/ui/views/CursorIconButton.java
@@ -20,7 +20,6 @@ package com.waz.zclient.ui.views;
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.GradientDrawable;
 import android.graphics.drawable.StateListDrawable;
 import android.util.AttributeSet;
